+++ .env.sample
@@ -14,3 +14,7 @@ REDIS_PORT=6379
 
 ADMIN_LOGIN=admin
 ADMIN_PASSWORD=passwd
+
+JWT_SECRET_KEY=super-secret
+JWT_ACCESS_TOKEN_EXPIRES=60
+JWT_REFRESH_TOKEN_EXPIRES=2592000
