+++ src/app/api/admin/__init__.py
@@ -1,1 +1,5 @@
-from app.api.admin.namespaces import namespace
+from flask_restplus import Namespace
+
+namespace = Namespace("admin", path="/admin", description="Admin API operations")
+
+from app.api.admin import roles, users
