--- src/app/api/admin/namespaces.py
@@ -1,3 +1,0 @@
-from flask_restplus import Namespace
-
-namespace = Namespace("admin", path="/admin", description="Auth admin API")
