+++ src/app/api/admin/schemas.py
@@ -1,0 +1,10 @@
+from flask_restplus import fields
+
+from app.api.admin import namespace
+
+admin_role_schema = namespace.model("Role", {
+    "id": fields.Integer(),
+    "name": fields.String(),
+    "description": fields.String(),
+    "is_active": fields.Boolean(),
+})
