+++ src/app/api/v1/__init__.py
@@ -1,1 +1,5 @@
-from app.api.v1 import auth
+from flask_restplus import Namespace
+
+namespace = Namespace("auth v1", path="/api/v1", description="Auth API v1 operations")
+
+from app.api.v1 import accounts, users
