--- src/app/api/v1/auth.py
@@ -1,4 +1,0 @@
-from flask_restplus import Namespace
-
-
-namespace = Namespace("auth", path="/api", description="Auth API operations")
